Error handling and minor changes
Added error handling of the user input, the program writes an error
message to stderr on error and exits the program. Also added error
handling to map_init where the program exits if an error occurs.

Added static to the functions that are not main as they shouldn't be
called from outside main.c.

Added a maximum value for the height on the map I plan to use later.