Better error handling
Made a function print_err_exit that takes the name of a function where
an error occurred and an error number from errno.h. It then prints the
function name, the error number and a string version of the error.

Changed my current error handling with custom error string to use the
new error function.

Changed the call of map_init to use the new interface and
map_save_as_png to use the new interface.