Made map have a single side length
When creating a map with a large length and a small width the map was
very badly made. I therefore changed width and length to a single number
side_len. This means that you can now only create a map that is the same
length on all sides.

To allow for a user to create a map that is not the same length on both
sides I have made map_save_as_png take the size of the image that should
be created. This means that the map is made correctly as it is made with
the same length on both sides, but it can later be saved as the size
requested by the user.

I also changed some occurrences of int to size_t to make sure that the
numbers passed are not negative.