Added error handling and removed length
Added error handling to the function map_save_as_png and changed the
function to take a height and a width of the picture the user wants.

Added function to allocate the height array and a function to handle
half allocated arrays.

Changed map_init to no longer take both a width and a length and
instead take a side length. This is done because the map didn't handle
differences in width and length very gracefully.