Exposed that maps are only sizes power of 2
Made the function map_init take a number that is raised to the power of
two to make the map side length instead of taking an actual side length.
This is done to make sure that a user is aware that maps can only have
sizes that are one greater than the power of two.