Added new function to find new map value
The new function takes into account the previous value of the map and
the size of the map piece that it is choosing a value for. The greater
the map piece, the greater deviation there might be from the given
offset value. The function also uses the maps min and max to make sure
that no value smaller or greater than these are inserted.