Gave square and diamond responsibility for new val
Gave the functions square and diamond the responsibility for computing a
new value for the map. This was done to simplify the divide function.
The new way of computing the new value uses only the map and the size of
the map piece that is currently considered. I therefore don't have to
parse along the random_range.