Replaced random range with roughness
Instead of choosing the maximum amount a map can change between two
points a user now chooses a roughness parameter between 0 and 1. This
value is used to scale all changes to the map, so small values of
roughness gives small changes and large values gives large changes. This
results in a flat and hilly landscape.