Average function takes into account invalid values
In the diamond phase of the map updating some of the points to take the
mean of might lie outside the map itself. This is reflected by its
arguments being less than 0.

Before I had simply set the map_get_height method to return 0, but this
resulted in a lesser result than intended as the average function still
divided by 4 and not a smaller number. Now the average function counts
the number of invalid values and only divide by the number of correct
values.