Scale randomness from stepsize
Before the randomness would just be scaled by the constant 0.1. However
it makes much more sense for the scaling to be based on the stepsize.
Over the whole map it should be possible for any change in height
however over smaller distances only smaller changes should be allowed.
The new scaling scales with 1.0 over the whole map 0.5 over half the map
and so on. That gives much nicer maps overall.